+++ helpers/server/fetch-categories.ts
@@ -6,12 +6,16 @@ import { PaginatedResult } from 'shared/types/result';
 import { sdk } from 'sdk';
 import getServerOptions from './get-server-options';
 
-const fetchCategories = cache(({ format = 'flat', limit = 500 }: { format?: 'tree' | 'flat'; limit?: number } = {}) => {
+const memoizedFetchCategories = cache((format = 'flat', limit = 500) => {
   return sdk.callAction<PaginatedResult<Category>>({
     actionName: 'product/queryCategories',
     query: { format, limit },
     ...getServerOptions(),
   });
+});
+
+const fetchCategories = cache(({ format = 'flat', limit = 500 }: { format?: 'tree' | 'flat'; limit?: number } = {}) => {
+  return memoizedFetchCategories(format, limit);
 });
 
 export default fetchCategories;